backend/requirements.txt
Django==3.2.12
djangorestframework==3.12.4
psycopg2==2.8
django-cors-headers==3.8.0
pytest-django==4.8.0  

